Logging on three phases, modified quadrature routine
The existing quadrature routine is way too sensitive to both timing and
line noise: if it skipped a single crossing (e.g. bouncing noise) the
count would be corrupted by 1 _invariably_. This (yet to be tested)
routine should improve immunity against bouncing.

Logging on three phases is still problematic: the timing of phase 2
suggests a false positive AC trigger from transients at the posedge of
PWM.

Coming up: experimenting with resistor divider networks to filter out
oscillations of Vdd. The range of divider coefficients is slim: we need
two resistor dividers, and the ratio of the two must be within
R_{DS,P,on} / (R_{DS,N,on} + R_{motor}), which is usually 3-5%.
Otherwise, the divided values are outside of the dynamic range of the
signal from the output nodes of the motor driver.