Loop closed
The encoder is now part of the action. The inverter and main code
step through the phases now under the beat of the encoder ticking.
Preliminary tests show that for the 1600CPR encoder currently
attached to the EMax motor, the interrupt routine for bidirectional
decoding was too slow, so the phase drift widened until the motor
was inefficient enough to stop accelerating. Replacing the routine
with a more... efficient one (just listen on one channel and know a
priori which way it's going), it looks like the encoder itself becomes
the limiting factor. The motor can reach speeds of ~4.5kRPM which nears
the limit of the encoder rating. Of course, all this is very
encouraging becaause the encoder is two orders of magnitude too fine;
we'll likely settle for something closer to 50CPR, and I'm hopeful
to reach beyond the kV of the motor with a faster processor and
off-chip decoding.

Using angle estimation gives either fantastic phase matching or
terrible phase matching, not much in between, so I am assuming there's
a sign inconsistency sometimes. The direction sensing in the
FirstEncoderTick phase is the prime suspect, since it's oversensitive.